use oom_score_adj instead of deprecated oom_adj
We are trying to get rid of oom_adj in the kernel. It has been
deprecated since 2.6.36. Let's use oom_score_adj preferably
and fallback to oom_adj when the former is not present (if
used on older systems). As the input expectes only oom_adj
values we have to recalculate it to oom_score_adj scale
(this is taken from fs/proc/base.c:oom_adj_write from the
kernel source tree).

Signed-off-by: Michal Hocko <mstsxfx@gmail.com>